rename module pykeen.nn.emb -> pykeen.nn.representation
