+++ src/pykeen/models/multimodal/complex_literal.py
@@ -11,7 +11,7 @@ from .base import LiteralModel
 from ...constants import DEFAULT_DROPOUT_HPO_RANGE, DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
 from ...losses import BCEWithLogitsLoss, Loss
 from ...nn.combinations import ComplExLiteralCombination
-from ...nn.emb import EmbeddingSpecification
+from ...nn.representation import EmbeddingSpecification
 from ...nn.modules import ComplExInteraction, LiteralInteraction
 from ...triples import TriplesNumericLiteralsFactory
