+++ src/pykeen/models/multimodal/distmult_literal.py
@@ -9,7 +9,7 @@ import torch.nn as nn
 from .base import LiteralModel
 from ...constants import DEFAULT_DROPOUT_HPO_RANGE, DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
 from ...nn.combinations import DistMultCombination
-from ...nn.emb import EmbeddingSpecification
+from ...nn.representation import EmbeddingSpecification
 from ...nn.modules import DistMultInteraction, LiteralInteraction
 from ...triples import TriplesNumericLiteralsFactory
