+++ src/pykeen/models/resolve.py
@@ -61,7 +61,7 @@ import logging
 from typing import Any, Mapping, Optional, Sequence, Tuple, Type, Union
 
 from .nbase import EmbeddingSpecificationHint, ERModel
-from ..nn.emb import EmbeddingSpecification, RepresentationModule
+from ..nn.representation import EmbeddingSpecification, RepresentationModule
 from ..nn.modules import Interaction, interaction_resolver
 from ..typing import HeadRepresentation, RelationRepresentation, TailRepresentation
