+++ src/pykeen/models/unimodal/auto_sf.py
@@ -6,7 +6,7 @@ from typing import Any, ClassVar, Mapping, Optional, Sequence, Tuple
 
 from ..nbase import ERModel
 from ...constants import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
-from ...nn.emb import EmbeddingSpecification
+from ...nn.representation import EmbeddingSpecification
 from ...nn.modules import AutoSFInteraction
 from ...typing import Sign
