+++ src/pykeen/models/unimodal/compgcn.py
@@ -8,7 +8,7 @@ import torch
 from class_resolver import Hint
 
 from ..nbase import ERModel
-from ...nn.emb import CombinedCompGCNRepresentations, EmbeddingSpecification
+from ...nn.representation import CombinedCompGCNRepresentations, EmbeddingSpecification
 from ...nn.modules import DistMultInteraction, Interaction
 from ...triples import CoreTriplesFactory
 from ...typing import RelationRepresentation
