+++ src/pykeen/models/unimodal/cp.py
@@ -8,7 +8,7 @@ import torch
 
 from ..nbase import ERModel
 from ...constants import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
-from ...nn.emb import EmbeddingSpecification
+from ...nn.representation import EmbeddingSpecification
 from ...nn.modules import CPInteraction
 from ...typing import Hint, InductiveMode, Initializer, Normalizer
