+++ src/pykeen/models/unimodal/kg2e.py
@@ -10,7 +10,7 @@ from torch.nn.init import uniform_
 
 from ..nbase import ERModel
 from ...constants import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
-from ...nn.emb import EmbeddingSpecification
+from ...nn.representation import EmbeddingSpecification
 from ...nn.modules import KG2EInteraction
 from ...typing import Constrainer, Hint, Initializer
 from ...utils import clamp_norm
