+++ src/pykeen/models/unimodal/rotate.py
@@ -9,7 +9,7 @@ import torch.autograd
 from torch import linalg
 
 from ..base import EntityRelationEmbeddingModel
-from ...nn.emb import EmbeddingSpecification
+from ...nn.representation import EmbeddingSpecification
 from ...nn.init import init_phases, xavier_uniform_
 from ...typing import Constrainer, Hint, Initializer
 from ...utils import complex_normalize
