+++ src/pykeen/models/unimodal/simple.py
@@ -9,7 +9,7 @@ from class_resolver import OptionalKwargs
 from ..nbase import ERModel
 from ...constants import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
 from ...losses import Loss, SoftplusLoss
-from ...nn.emb import EmbeddingSpecification
+from ...nn.representation import EmbeddingSpecification
 from ...nn.modules import SimplEInteraction
 from ...regularizers import PowerSumRegularizer, Regularizer, regularizer_resolver
 from ...typing import Hint, Initializer
