+++ src/pykeen/models/unimodal/trans_h.py
@@ -11,7 +11,7 @@ from torch.nn.init import uniform_
 
 from ..base import EntityRelationEmbeddingModel
 from ...constants import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
-from ...nn.emb import Embedding, EmbeddingSpecification
+from ...nn.representation import Embedding, EmbeddingSpecification
 from ...regularizers import Regularizer, TransHRegularizer
 from ...typing import Hint, Initializer
