+++ src/pykeen/models/unimodal/trans_r.py
@@ -12,7 +12,7 @@ from torch import linalg
 
 from ..base import EntityRelationEmbeddingModel
 from ...constants import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
-from ...nn.emb import Embedding, EmbeddingSpecification
+from ...nn.representation import Embedding, EmbeddingSpecification
 from ...nn.init import xavier_uniform_, xavier_uniform_norm_
 from ...typing import Constrainer, Hint, Initializer
 from ...utils import clamp_norm
