+++ src/pykeen/models/unimodal/unstructured_model.py
@@ -6,7 +6,7 @@ from typing import Any, ClassVar, Mapping
 
 from ..nbase import ERModel
 from ...constants import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
-from ...nn.emb import EmbeddingSpecification
+from ...nn.representation import EmbeddingSpecification
 from ...nn.init import xavier_normal_
 from ...nn.modules import UMInteraction
 from ...typing import Hint, Initializer
