+++ tests/cases.py
@@ -39,7 +39,7 @@ from torch.nn import functional
 from torch.optim import SGD, Adagrad
 
 import pykeen.models
-import pykeen.nn.emb
+import pykeen.nn.representation
 import pykeen.nn.message_passing
 import pykeen.nn.node_piece
 import pykeen.nn.weighting
@@ -53,7 +53,7 @@ from pykeen.losses import Loss, PairwiseLoss, PointwiseLoss, SetwiseLoss, Unsupp
 from pykeen.models import RESCAL, EntityRelationEmbeddingModel, Model, TransE
 from pykeen.models.cli import build_cli_from_cls
 from pykeen.models.nbase import ERModel
-from pykeen.nn.emb import RepresentationModule
+from pykeen.nn.representation import RepresentationModule
 from pykeen.nn.modules import FunctionalInteraction, Interaction, LiteralInteraction
 from pykeen.optimizers import optimizer_resolver
 from pykeen.pipeline import pipeline
